+++ Changelog.md
@@ -11,6 +11,7 @@
 * Fix `badAriaAttributeValue` not correctly handling decimal values (#182).
 * Work around null pointer exception caused by closure compiler issue (#183).
 * Add a special case to handle color `"transparent"` to fix #180.
+* Fix false positives on elements with no role for Unsupported ARIA Attribute rule (#199).
 
 ## 2.8.0 - 2015-07-24